+++ src/api_params.rs
@@ -101,7 +101,7 @@ pub struct SetWebhookParams {
     pub ip_address: Option<String>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_connections: Option<isize>,
+    pub max_connections: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub allowed_updates: Option<Vec<String>>,
@@ -412,7 +412,7 @@ pub struct SendVideoNoteParams {
     pub duration: Option<u32>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub length: Option<isize>,
+    pub length: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub thumb: Option<FileEnum>,
@@ -458,13 +458,13 @@ pub struct SendLocationParams {
     pub horizontal_accuracy: Option<f64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub live_period: Option<isize>,
+    pub live_period: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub heading: Option<isize>,
+    pub heading: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub proximity_alert_radius: Option<isize>,
+    pub proximity_alert_radius: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub disable_notification: Option<bool>,
@@ -498,10 +498,10 @@ pub struct EditMessageLiveLocationParams {
     pub horizontal_accuracy: Option<f64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub heading: Option<isize>,
+    pub heading: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub proximity_alert_radius: Option<isize>,
+    pub proximity_alert_radius: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub reply_markup: Option<InlineKeyboardMarkup>,
@@ -604,7 +604,7 @@ pub struct SendPollParams {
     pub allows_multiple_answers: Option<bool>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub correct_option_id: Option<isize>,
+    pub correct_option_id: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub explanation: Option<String>,
@@ -616,7 +616,7 @@ pub struct SendPollParams {
     pub explanation_entities: Option<Vec<MessageEntity>>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub open_period: Option<isize>,
+    pub open_period: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub close_date: Option<u64>,
@@ -910,7 +910,7 @@ pub struct AnswerCallbackQueryParams {
     pub url: Option<String>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub cache_time: Option<isize>,
+    pub cache_time: Option<i64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -1093,7 +1093,7 @@ pub struct AddStickerToSetParams {
 pub struct SetStickerPositionInSetParams {
     pub sticker: String,
 
-    pub position: isize,
+    pub position: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -1118,7 +1118,7 @@ pub struct AnswerInlineQueryParams {
     pub results: Vec<InlineQueryResult>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub cache_time: Option<isize>,
+    pub cache_time: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub is_personal: Option<bool>,
@@ -1150,10 +1150,10 @@ pub struct SendInvoiceParams {
     pub prices: Vec<LabeledPrice>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_tip_amount: Option<isize>,
+    pub max_tip_amount: Option<i64>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub suggested_tip_amounts: Option<Vec<isize>>,
+    pub suggested_tip_amounts: Option<Vec<i64>>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub start_parameter: Option<String>,
@@ -1260,7 +1260,7 @@ pub struct SendGameParams {
 pub struct SetGameScoreParams {
     pub user_id: u64,
 
-    pub score: isize,
+    pub score: i64,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub force: Option<bool>,
@@ -1489,7 +1489,7 @@ impl SetWebhookParams {
         self.ip_address = ip_address;
     }
 
-    pub fn set_max_connections(&mut self, max_connections: Option<isize>) {
+    pub fn set_max_connections(&mut self, max_connections: Option<i64>) {
         self.max_connections = max_connections;
     }
 
@@ -1513,7 +1513,7 @@ impl SetWebhookParams {
         self.ip_address.clone()
     }
 
-    pub fn max_connections(&self) -> Option<isize> {
+    pub fn max_connections(&self) -> Option<i64> {
         self.max_connections
     }
 
@@ -2472,7 +2472,7 @@ impl SendVideoNoteParams {
         self.duration = duration;
     }
 
-    pub fn set_length(&mut self, length: Option<isize>) {
+    pub fn set_length(&mut self, length: Option<i64>) {
         self.length = length;
     }
 
@@ -2508,7 +2508,7 @@ impl SendVideoNoteParams {
         self.duration
     }
 
-    pub fn length(&self) -> Option<isize> {
+    pub fn length(&self) -> Option<i64> {
         self.length
     }
 
@@ -2618,15 +2618,15 @@ impl SendLocationParams {
         self.horizontal_accuracy = horizontal_accuracy;
     }
 
-    pub fn set_live_period(&mut self, live_period: Option<isize>) {
+    pub fn set_live_period(&mut self, live_period: Option<i64>) {
         self.live_period = live_period;
     }
 
-    pub fn set_heading(&mut self, heading: Option<isize>) {
+    pub fn set_heading(&mut self, heading: Option<i64>) {
         self.heading = heading;
     }
 
-    pub fn set_proximity_alert_radius(&mut self, proximity_alert_radius: Option<isize>) {
+    pub fn set_proximity_alert_radius(&mut self, proximity_alert_radius: Option<i64>) {
         self.proximity_alert_radius = proximity_alert_radius;
     }
 
@@ -2662,15 +2662,15 @@ impl SendLocationParams {
         self.horizontal_accuracy
     }
 
-    pub fn live_period(&self) -> Option<isize> {
+    pub fn live_period(&self) -> Option<i64> {
         self.live_period
     }
 
-    pub fn heading(&self) -> Option<isize> {
+    pub fn heading(&self) -> Option<i64> {
         self.heading
     }
 
-    pub fn proximity_alert_radius(&self) -> Option<isize> {
+    pub fn proximity_alert_radius(&self) -> Option<i64> {
         self.proximity_alert_radius
     }
 
@@ -2730,11 +2730,11 @@ impl EditMessageLiveLocationParams {
         self.horizontal_accuracy = horizontal_accuracy;
     }
 
-    pub fn set_heading(&mut self, heading: Option<isize>) {
+    pub fn set_heading(&mut self, heading: Option<i64>) {
         self.heading = heading;
     }
 
-    pub fn set_proximity_alert_radius(&mut self, proximity_alert_radius: Option<isize>) {
+    pub fn set_proximity_alert_radius(&mut self, proximity_alert_radius: Option<i64>) {
         self.proximity_alert_radius = proximity_alert_radius;
     }
 
@@ -2766,11 +2766,11 @@ impl EditMessageLiveLocationParams {
         self.horizontal_accuracy
     }
 
-    pub fn heading(&self) -> Option<isize> {
+    pub fn heading(&self) -> Option<i64> {
         self.heading
     }
 
-    pub fn proximity_alert_radius(&self) -> Option<isize> {
+    pub fn proximity_alert_radius(&self) -> Option<i64> {
         self.proximity_alert_radius
     }
 
@@ -3087,7 +3087,7 @@ impl SendPollParams {
         self.allows_multiple_answers = allows_multiple_answers;
     }
 
-    pub fn set_correct_option_id(&mut self, correct_option_id: Option<isize>) {
+    pub fn set_correct_option_id(&mut self, correct_option_id: Option<i64>) {
         self.correct_option_id = correct_option_id;
     }
 
@@ -3103,7 +3103,7 @@ impl SendPollParams {
         self.explanation_entities = explanation_entities;
     }
 
-    pub fn set_open_period(&mut self, open_period: Option<isize>) {
+    pub fn set_open_period(&mut self, open_period: Option<i64>) {
         self.open_period = open_period;
     }
 
@@ -3155,7 +3155,7 @@ impl SendPollParams {
         self.allows_multiple_answers
     }
 
-    pub fn correct_option_id(&self) -> Option<isize> {
+    pub fn correct_option_id(&self) -> Option<i64> {
         self.correct_option_id
     }
 
@@ -3171,7 +3171,7 @@ impl SendPollParams {
         self.explanation_entities.clone()
     }
 
-    pub fn open_period(&self) -> Option<isize> {
+    pub fn open_period(&self) -> Option<i64> {
         self.open_period
     }
 
@@ -4050,7 +4050,7 @@ impl AnswerCallbackQueryParams {
         self.url = url;
     }
 
-    pub fn set_cache_time(&mut self, cache_time: Option<isize>) {
+    pub fn set_cache_time(&mut self, cache_time: Option<i64>) {
         self.cache_time = cache_time;
     }
 
@@ -4070,7 +4070,7 @@ impl AnswerCallbackQueryParams {
         self.url.clone()
     }
 
-    pub fn cache_time(&self) -> Option<isize> {
+    pub fn cache_time(&self) -> Option<i64> {
         self.cache_time
     }
 }
@@ -4633,7 +4633,7 @@ impl AddStickerToSetParams {
 }
 
 impl SetStickerPositionInSetParams {
-    pub fn new(sticker: String, position: isize) -> Self {
+    pub fn new(sticker: String, position: i64) -> Self {
         Self { sticker, position }
     }
 
@@ -4641,7 +4641,7 @@ impl SetStickerPositionInSetParams {
         self.sticker = sticker;
     }
 
-    pub fn set_position(&mut self, position: isize) {
+    pub fn set_position(&mut self, position: i64) {
         self.position = position;
     }
 
@@ -4649,7 +4649,7 @@ impl SetStickerPositionInSetParams {
         self.sticker.clone()
     }
 
-    pub fn position(&self) -> isize {
+    pub fn position(&self) -> i64 {
         self.position
     }
 }
@@ -4723,7 +4723,7 @@ impl AnswerInlineQueryParams {
         self.results = results;
     }
 
-    pub fn set_cache_time(&mut self, cache_time: Option<isize>) {
+    pub fn set_cache_time(&mut self, cache_time: Option<i64>) {
         self.cache_time = cache_time;
     }
 
@@ -4751,7 +4751,7 @@ impl AnswerInlineQueryParams {
         self.results.clone()
     }
 
-    pub fn cache_time(&self) -> Option<isize> {
+    pub fn cache_time(&self) -> Option<i64> {
         self.cache_time
     }
 
@@ -4848,11 +4848,11 @@ impl SendInvoiceParams {
         self.provider_data = provider_data;
     }
 
-    pub fn set_max_tip_amount(&mut self, max_tip_amount: Option<isize>) {
+    pub fn set_max_tip_amount(&mut self, max_tip_amount: Option<i64>) {
         self.max_tip_amount = max_tip_amount;
     }
 
-    pub fn set_suggested_tip_amounts(&mut self, suggested_tip_amounts: Option<Vec<isize>>) {
+    pub fn set_suggested_tip_amounts(&mut self, suggested_tip_amounts: Option<Vec<i64>>) {
         self.suggested_tip_amounts = suggested_tip_amounts;
     }
 
@@ -4951,11 +4951,11 @@ impl SendInvoiceParams {
         self.prices.clone()
     }
 
-    pub fn max_tip_amount(&self) -> Option<isize> {
+    pub fn max_tip_amount(&self) -> Option<i64> {
         self.max_tip_amount
     }
 
-    pub fn suggested_tip_amounts(&self) -> Option<Vec<isize>> {
+    pub fn suggested_tip_amounts(&self) -> Option<Vec<i64>> {
         self.suggested_tip_amounts.clone()
     }
 
@@ -5185,7 +5185,7 @@ impl SendGameParams {
 }
 
 impl SetGameScoreParams {
-    pub fn new(user_id: u64, score: isize) -> Self {
+    pub fn new(user_id: u64, score: i64) -> Self {
         Self {
             user_id,
             score,
@@ -5201,7 +5201,7 @@ impl SetGameScoreParams {
         self.user_id = user_id;
     }
 
-    pub fn set_score(&mut self, score: isize) {
+    pub fn set_score(&mut self, score: i64) {
         self.score = score;
     }
 
@@ -5229,7 +5229,7 @@ impl SetGameScoreParams {
         self.user_id
     }
 
-    pub fn score(&self) -> isize {
+    pub fn score(&self) -> i64 {
         self.score
     }
